apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
supabase==2.12.0
pytest==8.3.4
pytest-asyncio==0.25.3
